examples/todo: Implement as TodoFay
Re-implement the logic of the TodoCtrl angular controller using a Fay
module. The controller object itself is still javascript, but written
in a way that should be easy to generate.

Caveats:
  * setting the initial state is implemented as a function, called with
a dummy argument:

  state = new Strict.TodoFay.initialState(7);

It would be more natural to instead write:

  state = Strict.TodoFay.initialState;

However this generates:

  TS { text: ..., done: false }

rather than:

  Object { instance: 'TS', text: ..., done: false }

When generating the TS{..} form, the page shows 2 empty items. Their
text is shown after a function (such as addTodo()) is called, which
changes the state into the Object{..} form.